Move CTieCondition traits
