+++ include/football/CTieCondition.h
@@ -1,31 +1,13 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CTieCondition.h"
 
 #include "types/CTieCondition.h"
 
 #include <optional>
 
-namespace futsim
-{
-
-namespace football
-{
-class CTieCondition;
-}
-
-template <> struct json_traits<football::CTieCondition>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Tie condition";
-	//! JSON key for the \copybrief football::CTieCondition::mHomeTeamLead
-	static inline constexpr std::string_view HOME_TEAM_LEAD_KEY = "Home team lead";
-	//! JSON key for the \copybrief football::CTieCondition::mHomeTeamGoals
-	static inline constexpr std::string_view HOME_TEAM_GOALS_KEY = "Home team goals";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -94,6 +76,4 @@ private:
 	optional_goal_count mHomeTeamGoals;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
